Move tests for the setup module
